+++ R/mesa-package.R
@@ -1,0 +1,3 @@
+#' @import qsea
+#' @keywords internal
+"_PACKAGE"